rename algorithm classes and remove state re #326
