+++ mslice/cli/_mslice_commmands.py
@@ -18,9 +18,9 @@ from mslice.presenters.slice_plotter_presenter import Axis as _Axis
 from mslice.models.projection.powder.mantid_projection_calculator import MantidProjectionCalculator as _MantidProjectionCalculator
 # Slicing
 from mslice.models.slice.matplotlib_slice_plotter import MatplotlibSlicePlotter as _MatplotlibSlicePlotter
-from mslice.models.slice.mantid_slice_algorithm import MantidSliceAlgorithm as _MantidSliceAlgorithm
+import mslice.models.slice.slice_functions as _SLICE_ALGORITHM
 # Cutting
-from mslice.models.cut.mantid_cut_algorithm import MantidCutAlgorithm as _MantidCutAlgorithm
+import mslice.models.cut.cut_functions as _CUT_ALGORITHM
 from mslice.models.cut.matplotlib_cut_plotter import MatplotlibCutPlotter
 
 # -----------------------------------------------------------------------------
@@ -28,9 +28,7 @@ from mslice.models.cut.matplotlib_cut_plotter import MatplotlibCutPlotter
 # -----------------------------------------------------------------------------
 
 _POWDER_PROJECTION_MODEL = _MantidProjectionCalculator()
-_SLICE_ALGORITHM = _MantidSliceAlgorithm()
-_SLICE_MODEL = _MatplotlibSlicePlotter(_SLICE_ALGORITHM)
-_CUT_ALGORITHM = _MantidCutAlgorithm()
+_SLICE_MODEL = _MatplotlibSlicePlotter()
 _CUT_PLOTTER = MatplotlibCutPlotter(_CUT_ALGORITHM)
 
 # -----------------------------------------------------------------------------
