--- mslice/models/slice/slice_algorithm.py
@@ -1,10 +1,0 @@
-
-
-class SliceAlgorithm(object):
-
-    def compute_slice(self, selected_workspace, x_axis, y_axis, norm_to_one):
-        """Return a numpy masked matrix ready for plotting """
-        pass
-
-    def get_available_axis(self, workspace):
-        pass
